+++ eventbridge-20200401/alibabacloud_eventbridge20200401/models/_create_event_streaming_request.py
@@ -1676,6 +1676,9 @@ class CreateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
         sasl_user: str = None,
         security_group_id: str = None,
         security_protocol: str = None,
+        ssl_key_password: str = None,
+        ssl_keystore_certificate_chain: str = None,
+        ssl_keystore_key: main_models.CreateEventStreamingRequestSourceSourceApacheKafkaParametersSslKeystoreKey = None,
         ssl_truststore_certificates: str = None,
         topic: str = None,
         v_switch_ids: str = None,
@@ -1691,6 +1694,9 @@ class CreateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
         self.sasl_user = sasl_user
         self.security_group_id = security_group_id
         self.security_protocol = security_protocol
+        self.ssl_key_password = ssl_key_password
+        self.ssl_keystore_certificate_chain = ssl_keystore_certificate_chain
+        self.ssl_keystore_key = ssl_keystore_key
         self.ssl_truststore_certificates = ssl_truststore_certificates
         self.topic = topic
         self.v_switch_ids = v_switch_ids
@@ -1698,7 +1704,8 @@ class CreateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
         self.vpc_id = vpc_id
 
     def validate(self):
-        pass
+        if self.ssl_keystore_key:
+            self.ssl_keystore_key.validate()
 
     def to_map(self):
         result = dict()
@@ -1731,6 +1738,15 @@ class CreateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
 
         if self.security_protocol is not None:
             result['SecurityProtocol'] = self.security_protocol
+
+        if self.ssl_key_password is not None:
+            result['SslKeyPassword'] = self.ssl_key_password
+
+        if self.ssl_keystore_certificate_chain is not None:
+            result['SslKeystoreCertificateChain'] = self.ssl_keystore_certificate_chain
+
+        if self.ssl_keystore_key is not None:
+            result['SslKeystoreKey'] = self.ssl_keystore_key.to_map()
 
         if self.ssl_truststore_certificates is not None:
             result['SslTruststoreCertificates'] = self.ssl_truststore_certificates
@@ -1778,6 +1794,16 @@ class CreateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
         if m.get('SecurityProtocol') is not None:
             self.security_protocol = m.get('SecurityProtocol')
 
+        if m.get('SslKeyPassword') is not None:
+            self.ssl_key_password = m.get('SslKeyPassword')
+
+        if m.get('SslKeystoreCertificateChain') is not None:
+            self.ssl_keystore_certificate_chain = m.get('SslKeystoreCertificateChain')
+
+        if m.get('SslKeystoreKey') is not None:
+            temp_model = main_models.CreateEventStreamingRequestSourceSourceApacheKafkaParametersSslKeystoreKey()
+            self.ssl_keystore_key = temp_model.from_map(m.get('SslKeystoreKey'))
+
         if m.get('SslTruststoreCertificates') is not None:
             self.ssl_truststore_certificates = m.get('SslTruststoreCertificates')
 
@@ -1792,6 +1818,41 @@ class CreateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
 
         if m.get('VpcId') is not None:
             self.vpc_id = m.get('VpcId')
+
+        return self
+
+class CreateEventStreamingRequestSourceSourceApacheKafkaParametersSslKeystoreKey(DaraModel):
+    def __init__(
+        self,
+        kms_arn: str = None,
+        kms_secret_value_key: str = None,
+    ):
+        self.kms_arn = kms_arn
+        self.kms_secret_value_key = kms_secret_value_key
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.kms_arn is not None:
+            result['KmsArn'] = self.kms_arn
+
+        if self.kms_secret_value_key is not None:
+            result['KmsSecretValueKey'] = self.kms_secret_value_key
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('KmsArn') is not None:
+            self.kms_arn = m.get('KmsArn')
+
+        if m.get('KmsSecretValueKey') is not None:
+            self.kms_secret_value_key = m.get('KmsSecretValueKey')
 
         return self
 
@@ -5328,6 +5389,7 @@ class CreateEventStreamingRequestSinkSinkKafkaParameters(DaraModel):
         self,
         acks: main_models.CreateEventStreamingRequestSinkSinkKafkaParametersAcks = None,
         compression_type: str = None,
+        dynamic_topic: main_models.CreateEventStreamingRequestSinkSinkKafkaParametersDynamicTopic = None,
         headers: main_models.CreateEventStreamingRequestSinkSinkKafkaParametersHeaders = None,
         instance_id: main_models.CreateEventStreamingRequestSinkSinkKafkaParametersInstanceId = None,
         key: main_models.CreateEventStreamingRequestSinkSinkKafkaParametersKey = None,
@@ -5341,6 +5403,7 @@ class CreateEventStreamingRequestSinkSinkKafkaParameters(DaraModel):
         # *   If you set this parameter to all, a response is returned when data is written to the leader and synchronized to the followers. In this mode, the performance is low, but the risk of data loss is also low. Data loss occurs if the leader and the followers fail at the same time.
         self.acks = acks
         self.compression_type = compression_type
+        self.dynamic_topic = dynamic_topic
         self.headers = headers
         # The ID of the ApsaraMQ for Kafka instance.
         self.instance_id = instance_id
@@ -5354,6 +5417,8 @@ class CreateEventStreamingRequestSinkSinkKafkaParameters(DaraModel):
     def validate(self):
         if self.acks:
             self.acks.validate()
+        if self.dynamic_topic:
+            self.dynamic_topic.validate()
         if self.headers:
             self.headers.validate()
         if self.instance_id:
@@ -5375,6 +5440,9 @@ class CreateEventStreamingRequestSinkSinkKafkaParameters(DaraModel):
 
         if self.compression_type is not None:
             result['CompressionType'] = self.compression_type
+
+        if self.dynamic_topic is not None:
+            result['DynamicTopic'] = self.dynamic_topic.to_map()
 
         if self.headers is not None:
             result['Headers'] = self.headers.to_map()
@@ -5401,6 +5469,10 @@ class CreateEventStreamingRequestSinkSinkKafkaParameters(DaraModel):
 
         if m.get('CompressionType') is not None:
             self.compression_type = m.get('CompressionType')
+
+        if m.get('DynamicTopic') is not None:
+            temp_model = main_models.CreateEventStreamingRequestSinkSinkKafkaParametersDynamicTopic()
+            self.dynamic_topic = temp_model.from_map(m.get('DynamicTopic'))
 
         if m.get('Headers') is not None:
             temp_model = main_models.CreateEventStreamingRequestSinkSinkKafkaParametersHeaders()
@@ -5609,6 +5681,49 @@ class CreateEventStreamingRequestSinkSinkKafkaParametersInstanceId(DaraModel):
         return self
 
 class CreateEventStreamingRequestSinkSinkKafkaParametersHeaders(DaraModel):
+    def __init__(
+        self,
+        form: str = None,
+        template: str = None,
+        value: str = None,
+    ):
+        self.form = form
+        self.template = template
+        self.value = value
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.form is not None:
+            result['Form'] = self.form
+
+        if self.template is not None:
+            result['Template'] = self.template
+
+        if self.value is not None:
+            result['Value'] = self.value
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Form') is not None:
+            self.form = m.get('Form')
+
+        if m.get('Template') is not None:
+            self.template = m.get('Template')
+
+        if m.get('Value') is not None:
+            self.value = m.get('Value')
+
+        return self
+
+class CreateEventStreamingRequestSinkSinkKafkaParametersDynamicTopic(DaraModel):
     def __init__(
         self,
         form: str = None,
@@ -8322,6 +8437,7 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         acks: str = None,
         bootstraps: str = None,
         compression_type: str = None,
+        dynamic_topic: main_models.CreateEventStreamingRequestSinkSinkApacheKafkaParametersDynamicTopic = None,
         headers: main_models.CreateEventStreamingRequestSinkSinkApacheKafkaParametersHeaders = None,
         key: main_models.CreateEventStreamingRequestSinkSinkApacheKafkaParametersKey = None,
         network_type: main_models.CreateEventStreamingRequestSinkSinkApacheKafkaParametersNetworkType = None,
@@ -8330,6 +8446,9 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         sasl_user: str = None,
         security_group_id: main_models.CreateEventStreamingRequestSinkSinkApacheKafkaParametersSecurityGroupId = None,
         security_protocol: str = None,
+        ssl_key_password: str = None,
+        ssl_keystore_certificate_chain: str = None,
+        ssl_keystore_key: main_models.CreateEventStreamingRequestSinkSinkApacheKafkaParametersSslKeystoreKey = None,
         ssl_truststore_certificates: str = None,
         topic: str = None,
         v_switch_ids: main_models.CreateEventStreamingRequestSinkSinkApacheKafkaParametersVSwitchIds = None,
@@ -8339,6 +8458,7 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         self.acks = acks
         self.bootstraps = bootstraps
         self.compression_type = compression_type
+        self.dynamic_topic = dynamic_topic
         self.headers = headers
         self.key = key
         self.network_type = network_type
@@ -8347,6 +8467,9 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         self.sasl_user = sasl_user
         self.security_group_id = security_group_id
         self.security_protocol = security_protocol
+        self.ssl_key_password = ssl_key_password
+        self.ssl_keystore_certificate_chain = ssl_keystore_certificate_chain
+        self.ssl_keystore_key = ssl_keystore_key
         self.ssl_truststore_certificates = ssl_truststore_certificates
         self.topic = topic
         self.v_switch_ids = v_switch_ids
@@ -8354,6 +8477,8 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         self.vpc_id = vpc_id
 
     def validate(self):
+        if self.dynamic_topic:
+            self.dynamic_topic.validate()
         if self.headers:
             self.headers.validate()
         if self.key:
@@ -8362,6 +8487,8 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
             self.network_type.validate()
         if self.security_group_id:
             self.security_group_id.validate()
+        if self.ssl_keystore_key:
+            self.ssl_keystore_key.validate()
         if self.v_switch_ids:
             self.v_switch_ids.validate()
         if self.value:
@@ -8382,6 +8509,9 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
 
         if self.compression_type is not None:
             result['CompressionType'] = self.compression_type
+
+        if self.dynamic_topic is not None:
+            result['DynamicTopic'] = self.dynamic_topic.to_map()
 
         if self.headers is not None:
             result['Headers'] = self.headers.to_map()
@@ -8406,6 +8536,15 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
 
         if self.security_protocol is not None:
             result['SecurityProtocol'] = self.security_protocol
+
+        if self.ssl_key_password is not None:
+            result['SslKeyPassword'] = self.ssl_key_password
+
+        if self.ssl_keystore_certificate_chain is not None:
+            result['SslKeystoreCertificateChain'] = self.ssl_keystore_certificate_chain
+
+        if self.ssl_keystore_key is not None:
+            result['SslKeystoreKey'] = self.ssl_keystore_key.to_map()
 
         if self.ssl_truststore_certificates is not None:
             result['SslTruststoreCertificates'] = self.ssl_truststore_certificates
@@ -8435,6 +8574,10 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         if m.get('CompressionType') is not None:
             self.compression_type = m.get('CompressionType')
 
+        if m.get('DynamicTopic') is not None:
+            temp_model = main_models.CreateEventStreamingRequestSinkSinkApacheKafkaParametersDynamicTopic()
+            self.dynamic_topic = temp_model.from_map(m.get('DynamicTopic'))
+
         if m.get('Headers') is not None:
             temp_model = main_models.CreateEventStreamingRequestSinkSinkApacheKafkaParametersHeaders()
             self.headers = temp_model.from_map(m.get('Headers'))
@@ -8462,6 +8605,16 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
 
         if m.get('SecurityProtocol') is not None:
             self.security_protocol = m.get('SecurityProtocol')
+
+        if m.get('SslKeyPassword') is not None:
+            self.ssl_key_password = m.get('SslKeyPassword')
+
+        if m.get('SslKeystoreCertificateChain') is not None:
+            self.ssl_keystore_certificate_chain = m.get('SslKeystoreCertificateChain')
+
+        if m.get('SslKeystoreKey') is not None:
+            temp_model = main_models.CreateEventStreamingRequestSinkSinkApacheKafkaParametersSslKeystoreKey()
+            self.ssl_keystore_key = temp_model.from_map(m.get('SslKeystoreKey'))
 
         if m.get('SslTruststoreCertificates') is not None:
             self.ssl_truststore_certificates = m.get('SslTruststoreCertificates')
@@ -8612,6 +8765,41 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParametersVSwitchIds(DaraMod
 
         return self
 
+class CreateEventStreamingRequestSinkSinkApacheKafkaParametersSslKeystoreKey(DaraModel):
+    def __init__(
+        self,
+        kms_arn: str = None,
+        kms_secret_value_key: str = None,
+    ):
+        self.kms_arn = kms_arn
+        self.kms_secret_value_key = kms_secret_value_key
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.kms_arn is not None:
+            result['KmsArn'] = self.kms_arn
+
+        if self.kms_secret_value_key is not None:
+            result['KmsSecretValueKey'] = self.kms_secret_value_key
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('KmsArn') is not None:
+            self.kms_arn = m.get('KmsArn')
+
+        if m.get('KmsSecretValueKey') is not None:
+            self.kms_secret_value_key = m.get('KmsSecretValueKey')
+
+        return self
+
 class CreateEventStreamingRequestSinkSinkApacheKafkaParametersSecurityGroupId(DaraModel):
     def __init__(
         self,
@@ -8742,6 +8930,49 @@ class CreateEventStreamingRequestSinkSinkApacheKafkaParametersKey(DaraModel):
         return self
 
 class CreateEventStreamingRequestSinkSinkApacheKafkaParametersHeaders(DaraModel):
+    def __init__(
+        self,
+        form: str = None,
+        template: str = None,
+        value: str = None,
+    ):
+        self.form = form
+        self.template = template
+        self.value = value
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.form is not None:
+            result['Form'] = self.form
+
+        if self.template is not None:
+            result['Template'] = self.template
+
+        if self.value is not None:
+            result['Value'] = self.value
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Form') is not None:
+            self.form = m.get('Form')
+
+        if m.get('Template') is not None:
+            self.template = m.get('Template')
+
+        if m.get('Value') is not None:
+            self.value = m.get('Value')
+
+        return self
+
+class CreateEventStreamingRequestSinkSinkApacheKafkaParametersDynamicTopic(DaraModel):
     def __init__(
         self,
         form: str = None,