+++ eventbridge-20200401/alibabacloud_eventbridge20200401/models/_update_event_streaming_request.py
@@ -1582,6 +1582,9 @@ class UpdateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
         sasl_user: str = None,
         security_group_id: str = None,
         security_protocol: str = None,
+        ssl_key_password: str = None,
+        ssl_keystore_certificate_chain: str = None,
+        ssl_keystore_key: main_models.UpdateEventStreamingRequestSourceSourceApacheKafkaParametersSslKeystoreKey = None,
         ssl_truststore_certificates: str = None,
         topic: str = None,
         v_switch_ids: str = None,
@@ -1597,6 +1600,9 @@ class UpdateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
         self.sasl_user = sasl_user
         self.security_group_id = security_group_id
         self.security_protocol = security_protocol
+        self.ssl_key_password = ssl_key_password
+        self.ssl_keystore_certificate_chain = ssl_keystore_certificate_chain
+        self.ssl_keystore_key = ssl_keystore_key
         self.ssl_truststore_certificates = ssl_truststore_certificates
         self.topic = topic
         self.v_switch_ids = v_switch_ids
@@ -1604,7 +1610,8 @@ class UpdateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
         self.vpc_id = vpc_id
 
     def validate(self):
-        pass
+        if self.ssl_keystore_key:
+            self.ssl_keystore_key.validate()
 
     def to_map(self):
         result = dict()
@@ -1637,6 +1644,15 @@ class UpdateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
 
         if self.security_protocol is not None:
             result['SecurityProtocol'] = self.security_protocol
+
+        if self.ssl_key_password is not None:
+            result['SslKeyPassword'] = self.ssl_key_password
+
+        if self.ssl_keystore_certificate_chain is not None:
+            result['SslKeystoreCertificateChain'] = self.ssl_keystore_certificate_chain
+
+        if self.ssl_keystore_key is not None:
+            result['SslKeystoreKey'] = self.ssl_keystore_key.to_map()
 
         if self.ssl_truststore_certificates is not None:
             result['SslTruststoreCertificates'] = self.ssl_truststore_certificates
@@ -1684,6 +1700,16 @@ class UpdateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
         if m.get('SecurityProtocol') is not None:
             self.security_protocol = m.get('SecurityProtocol')
 
+        if m.get('SslKeyPassword') is not None:
+            self.ssl_key_password = m.get('SslKeyPassword')
+
+        if m.get('SslKeystoreCertificateChain') is not None:
+            self.ssl_keystore_certificate_chain = m.get('SslKeystoreCertificateChain')
+
+        if m.get('SslKeystoreKey') is not None:
+            temp_model = main_models.UpdateEventStreamingRequestSourceSourceApacheKafkaParametersSslKeystoreKey()
+            self.ssl_keystore_key = temp_model.from_map(m.get('SslKeystoreKey'))
+
         if m.get('SslTruststoreCertificates') is not None:
             self.ssl_truststore_certificates = m.get('SslTruststoreCertificates')
 
@@ -1698,6 +1724,41 @@ class UpdateEventStreamingRequestSourceSourceApacheKafkaParameters(DaraModel):
 
         if m.get('VpcId') is not None:
             self.vpc_id = m.get('VpcId')
+
+        return self
+
+class UpdateEventStreamingRequestSourceSourceApacheKafkaParametersSslKeystoreKey(DaraModel):
+    def __init__(
+        self,
+        kms_arn: str = None,
+        kms_secret_value_key: str = None,
+    ):
+        self.kms_arn = kms_arn
+        self.kms_secret_value_key = kms_secret_value_key
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.kms_arn is not None:
+            result['KmsArn'] = self.kms_arn
+
+        if self.kms_secret_value_key is not None:
+            result['KmsSecretValueKey'] = self.kms_secret_value_key
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('KmsArn') is not None:
+            self.kms_arn = m.get('KmsArn')
+
+        if m.get('KmsSecretValueKey') is not None:
+            self.kms_secret_value_key = m.get('KmsSecretValueKey')
 
         return self
 
@@ -5341,6 +5402,7 @@ class UpdateEventStreamingRequestSinkSinkKafkaParameters(DaraModel):
         self,
         acks: main_models.UpdateEventStreamingRequestSinkSinkKafkaParametersAcks = None,
         compression_type: str = None,
+        dynamic_topic: main_models.UpdateEventStreamingRequestSinkSinkKafkaParametersDynamicTopic = None,
         headers: main_models.UpdateEventStreamingRequestSinkSinkKafkaParametersHeaders = None,
         instance_id: main_models.UpdateEventStreamingRequestSinkSinkKafkaParametersInstanceId = None,
         key: main_models.UpdateEventStreamingRequestSinkSinkKafkaParametersKey = None,
@@ -5354,6 +5416,7 @@ class UpdateEventStreamingRequestSinkSinkKafkaParameters(DaraModel):
         # *   If you set this parameter to all, a response is returned when data is written to the leader and synchronized to the followers. In this mode, the performance is low, but the risk of data loss is also low. Data loss occurs if the leader and the followers fail at the same time.
         self.acks = acks
         self.compression_type = compression_type
+        self.dynamic_topic = dynamic_topic
         # The metadata added to messages in the ApsaraMQ for Kafka instance.
         self.headers = headers
         # The ID of the ApsaraMQ for Kafka instance.
@@ -5368,6 +5431,8 @@ class UpdateEventStreamingRequestSinkSinkKafkaParameters(DaraModel):
     def validate(self):
         if self.acks:
             self.acks.validate()
+        if self.dynamic_topic:
+            self.dynamic_topic.validate()
         if self.headers:
             self.headers.validate()
         if self.instance_id:
@@ -5389,6 +5454,9 @@ class UpdateEventStreamingRequestSinkSinkKafkaParameters(DaraModel):
 
         if self.compression_type is not None:
             result['CompressionType'] = self.compression_type
+
+        if self.dynamic_topic is not None:
+            result['DynamicTopic'] = self.dynamic_topic.to_map()
 
         if self.headers is not None:
             result['Headers'] = self.headers.to_map()
@@ -5415,6 +5483,10 @@ class UpdateEventStreamingRequestSinkSinkKafkaParameters(DaraModel):
 
         if m.get('CompressionType') is not None:
             self.compression_type = m.get('CompressionType')
+
+        if m.get('DynamicTopic') is not None:
+            temp_model = main_models.UpdateEventStreamingRequestSinkSinkKafkaParametersDynamicTopic()
+            self.dynamic_topic = temp_model.from_map(m.get('DynamicTopic'))
 
         if m.get('Headers') is not None:
             temp_model = main_models.UpdateEventStreamingRequestSinkSinkKafkaParametersHeaders()
@@ -5639,6 +5711,49 @@ class UpdateEventStreamingRequestSinkSinkKafkaParametersHeaders(DaraModel):
         # The template based on which you want events to be transformed.
         self.template = template
         # The value before event transformation.
+        self.value = value
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.form is not None:
+            result['Form'] = self.form
+
+        if self.template is not None:
+            result['Template'] = self.template
+
+        if self.value is not None:
+            result['Value'] = self.value
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Form') is not None:
+            self.form = m.get('Form')
+
+        if m.get('Template') is not None:
+            self.template = m.get('Template')
+
+        if m.get('Value') is not None:
+            self.value = m.get('Value')
+
+        return self
+
+class UpdateEventStreamingRequestSinkSinkKafkaParametersDynamicTopic(DaraModel):
+    def __init__(
+        self,
+        form: str = None,
+        template: str = None,
+        value: str = None,
+    ):
+        self.form = form
+        self.template = template
         self.value = value
 
     def validate(self):
@@ -8537,6 +8652,7 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         acks: str = None,
         bootstraps: str = None,
         compression_type: str = None,
+        dynamic_topic: main_models.UpdateEventStreamingRequestSinkSinkApacheKafkaParametersDynamicTopic = None,
         headers: main_models.UpdateEventStreamingRequestSinkSinkApacheKafkaParametersHeaders = None,
         key: main_models.UpdateEventStreamingRequestSinkSinkApacheKafkaParametersKey = None,
         network_type: main_models.UpdateEventStreamingRequestSinkSinkApacheKafkaParametersNetworkType = None,
@@ -8545,6 +8661,9 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         sasl_user: str = None,
         security_group_id: main_models.UpdateEventStreamingRequestSinkSinkApacheKafkaParametersSecurityGroupId = None,
         security_protocol: str = None,
+        ssl_key_password: str = None,
+        ssl_keystore_certificate_chain: str = None,
+        ssl_keystore_key: main_models.UpdateEventStreamingRequestSinkSinkApacheKafkaParametersSslKeystoreKey = None,
         ssl_truststore_certificates: str = None,
         topic: str = None,
         v_switch_ids: main_models.UpdateEventStreamingRequestSinkSinkApacheKafkaParametersVSwitchIds = None,
@@ -8554,6 +8673,7 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         self.acks = acks
         self.bootstraps = bootstraps
         self.compression_type = compression_type
+        self.dynamic_topic = dynamic_topic
         self.headers = headers
         self.key = key
         self.network_type = network_type
@@ -8562,6 +8682,9 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         self.sasl_user = sasl_user
         self.security_group_id = security_group_id
         self.security_protocol = security_protocol
+        self.ssl_key_password = ssl_key_password
+        self.ssl_keystore_certificate_chain = ssl_keystore_certificate_chain
+        self.ssl_keystore_key = ssl_keystore_key
         self.ssl_truststore_certificates = ssl_truststore_certificates
         self.topic = topic
         self.v_switch_ids = v_switch_ids
@@ -8569,6 +8692,8 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         self.vpc_id = vpc_id
 
     def validate(self):
+        if self.dynamic_topic:
+            self.dynamic_topic.validate()
         if self.headers:
             self.headers.validate()
         if self.key:
@@ -8577,6 +8702,8 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
             self.network_type.validate()
         if self.security_group_id:
             self.security_group_id.validate()
+        if self.ssl_keystore_key:
+            self.ssl_keystore_key.validate()
         if self.v_switch_ids:
             self.v_switch_ids.validate()
         if self.value:
@@ -8597,6 +8724,9 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
 
         if self.compression_type is not None:
             result['CompressionType'] = self.compression_type
+
+        if self.dynamic_topic is not None:
+            result['DynamicTopic'] = self.dynamic_topic.to_map()
 
         if self.headers is not None:
             result['Headers'] = self.headers.to_map()
@@ -8621,6 +8751,15 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
 
         if self.security_protocol is not None:
             result['SecurityProtocol'] = self.security_protocol
+
+        if self.ssl_key_password is not None:
+            result['SslKeyPassword'] = self.ssl_key_password
+
+        if self.ssl_keystore_certificate_chain is not None:
+            result['SslKeystoreCertificateChain'] = self.ssl_keystore_certificate_chain
+
+        if self.ssl_keystore_key is not None:
+            result['SslKeystoreKey'] = self.ssl_keystore_key.to_map()
 
         if self.ssl_truststore_certificates is not None:
             result['SslTruststoreCertificates'] = self.ssl_truststore_certificates
@@ -8650,6 +8789,10 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
         if m.get('CompressionType') is not None:
             self.compression_type = m.get('CompressionType')
 
+        if m.get('DynamicTopic') is not None:
+            temp_model = main_models.UpdateEventStreamingRequestSinkSinkApacheKafkaParametersDynamicTopic()
+            self.dynamic_topic = temp_model.from_map(m.get('DynamicTopic'))
+
         if m.get('Headers') is not None:
             temp_model = main_models.UpdateEventStreamingRequestSinkSinkApacheKafkaParametersHeaders()
             self.headers = temp_model.from_map(m.get('Headers'))
@@ -8677,6 +8820,16 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParameters(DaraModel):
 
         if m.get('SecurityProtocol') is not None:
             self.security_protocol = m.get('SecurityProtocol')
+
+        if m.get('SslKeyPassword') is not None:
+            self.ssl_key_password = m.get('SslKeyPassword')
+
+        if m.get('SslKeystoreCertificateChain') is not None:
+            self.ssl_keystore_certificate_chain = m.get('SslKeystoreCertificateChain')
+
+        if m.get('SslKeystoreKey') is not None:
+            temp_model = main_models.UpdateEventStreamingRequestSinkSinkApacheKafkaParametersSslKeystoreKey()
+            self.ssl_keystore_key = temp_model.from_map(m.get('SslKeystoreKey'))
 
         if m.get('SslTruststoreCertificates') is not None:
             self.ssl_truststore_certificates = m.get('SslTruststoreCertificates')
@@ -8827,6 +8980,41 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParametersVSwitchIds(DaraMod
 
         return self
 
+class UpdateEventStreamingRequestSinkSinkApacheKafkaParametersSslKeystoreKey(DaraModel):
+    def __init__(
+        self,
+        kms_arn: str = None,
+        kms_secret_value_key: str = None,
+    ):
+        self.kms_arn = kms_arn
+        self.kms_secret_value_key = kms_secret_value_key
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.kms_arn is not None:
+            result['KmsArn'] = self.kms_arn
+
+        if self.kms_secret_value_key is not None:
+            result['KmsSecretValueKey'] = self.kms_secret_value_key
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('KmsArn') is not None:
+            self.kms_arn = m.get('KmsArn')
+
+        if m.get('KmsSecretValueKey') is not None:
+            self.kms_secret_value_key = m.get('KmsSecretValueKey')
+
+        return self
+
 class UpdateEventStreamingRequestSinkSinkApacheKafkaParametersSecurityGroupId(DaraModel):
     def __init__(
         self,
@@ -8957,6 +9145,49 @@ class UpdateEventStreamingRequestSinkSinkApacheKafkaParametersKey(DaraModel):
         return self
 
 class UpdateEventStreamingRequestSinkSinkApacheKafkaParametersHeaders(DaraModel):
+    def __init__(
+        self,
+        form: str = None,
+        template: str = None,
+        value: str = None,
+    ):
+        self.form = form
+        self.template = template
+        self.value = value
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.form is not None:
+            result['Form'] = self.form
+
+        if self.template is not None:
+            result['Template'] = self.template
+
+        if self.value is not None:
+            result['Value'] = self.value
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Form') is not None:
+            self.form = m.get('Form')
+
+        if m.get('Template') is not None:
+            self.template = m.get('Template')
+
+        if m.get('Value') is not None:
+            self.value = m.get('Value')
+
+        return self
+
+class UpdateEventStreamingRequestSinkSinkApacheKafkaParametersDynamicTopic(DaraModel):
     def __init__(
         self,
         form: str = None,